samples: nrf9160: modem_shell: mosh_print: shell ptr init added
assert() removed from mosh_print_no_format() in case if no
mosh_shell pointer. Initilaizing that instead of an assertion,
as it is a valid use case when mosh_print_no_format() is used
before mosh_fprintf.
This is a case at least when using with plain gnss system mode.
Signed-off-by: Jani Hirsimäki <jani.hirsimaki@nordicsemi.no>