subsys: net: l2_dect_nrp: sink: conn mgr usage: do not ignore ppp iface
Connection Manager usage changed to not to set ppp/prefix iface as ignore.
THis means that connection manager events will be received also for
ppp/prefix iface and in case of NET_EVENT_L4_IPV6_CONNECTED event,
it is received there because it is the 1st where global IPv6 address
is assigned to iface.

Signed-off-by: Jani Hirsimäki <jani.hirsimaki@nordicsemi.no>